refactor : #18 패키지 구조 수정
